Vector Equation Tail
I have scrapped the use of trigonometry to calculate the angle of tail
emission. I have added randomness and fixed the tail colour. It looks
good, however, the tail emission is now affected by the velocity of the
comet, therefore the angle of emission is incorrect as it approaches the
star.